Disambiguate call to std::pow.
git-svn-id: https://svn.dealii.org/trunk@18982 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/base/include/base/tensor.h
@@ -1436,13 +1436,13 @@ double determinant (const Tensor<2,dim> &t)
 	for (unsigned int j=0; j<dim-1; ++j)
 	  minor[i][j] = t[i][j<k ? j : j+1];
 
-      const double cofactor = std::pow (-1, k+1) *
+      const double cofactor = std::pow (-1., static_cast<double>(k+1)) *
 			      determinant (minor);
 
       det += t[dim-1][k] * cofactor;
     }
   
-  return std::pow (-1, dim) * det;
+  return std::pow (-1., static_cast<double>(dim)) * det;
 }
